+++ examples/client/wallet_info.py
