+++ examples/purchase/run_premium_giveaway.py
@@ -1,8 +1,8 @@
 """
-Example: purchase Telegram Stars for a user.
+Example: run a Telegram Premium giveaway for a channel.
 
-Amount must be an integer between 50 and 1 000 000.
-Set show_sender=False to send anonymously.
+winners must be an integer between 1 and 24 000.
+months (Premium duration per winner) must be 3, 6, or 12.
 """
 
 import asyncio
@@ -18,22 +18,29 @@ COOKIES = {
     "stel_ton_token": "YOUR_STEL_TON_TOKEN",
 }
 
-USERNAME = "@username"
-AMOUNT = 500  # 50–1 000 000 stars
+CHANNEL = "@channel"
+WINNERS = 10  # 1–24 000
+MONTHS = 3  # 3, 6 or 12
 
 
 async def main() -> None:
     async with FragmentClient(seed=SEED, api_key=API_KEY, cookies=COOKIES) as client:
         try:
-            result = await client.purchase_stars(USERNAME, amount=AMOUNT, show_sender=True)
+            result = await client.giveaway_premium(
+                CHANNEL, winners=WINNERS, months=MONTHS
+            )
         except UserNotFoundError:
-            print(f"User {USERNAME} was not found on fragment.com — check the username and try again.")
+            print(
+                f"Channel {CHANNEL} was not found on fragment.com — check the username and try again."
+            )
             return
         except ConfigurationError as e:
             print(f"Invalid argument: {e}")
             return
 
-    print(f"{result.amount} Stars successfully sent to {result.username} | tx: {result.transaction_id}")
+    print(
+        f"Premium giveaway created for {result.channel} — {result.winners} winner(s) × {result.amount} months each | tx: {result.transaction_id}"
+    )
 
 
 if __name__ == "__main__":
