+++ examples/purchase/run_stars_giveaway.py
@@ -26,9 +26,13 @@ AMOUNT = 1000  # 500–1 000 000 stars per winner
 async def main() -> None:
     async with FragmentClient(seed=SEED, api_key=API_KEY, cookies=COOKIES) as client:
         try:
-            result = await client.giveaway_stars(CHANNEL, winners=WINNERS, amount=AMOUNT)
+            result = await client.giveaway_stars(
+                CHANNEL, winners=WINNERS, amount=AMOUNT
+            )
         except UserNotFoundError:
-            print(f"Channel {CHANNEL} was not found on fragment.com — check the username and try again.")
+            print(
+                f"Channel {CHANNEL} was not found on fragment.com — check the username and try again."
+            )
             return
         except ConfigurationError as e:
             print(f"Invalid argument: {e}")
